Ошибка линтинга
