+++ src/components/login/login.tsx
@@ -3,8 +3,7 @@ import { Helmet } from 'react-helmet-async';
 import { Link } from 'react-router-dom';
 import { useAppDispatch } from '../../hooks';
 import { HelmetTitles } from '../../const';
-import { loginAction } from '../../store/user/user.slice';
-//import { loginAction } from '../../store/user/user.api-actions';
+import { loginAction } from '../../store/user/user.api-actions';
 
 function Login(): JSX.Element {
   const loginRef = useRef<HTMLInputElement | null>(null);
