+++ src/services/process-error-handle.ts
@@ -1,5 +1,6 @@
 import { store } from '../store';
-import { clearErrorAction, setError } from '../store/user/user.slice';
+import { clearErrorAction } from '../store/user/user.api-actions';
+import { setError } from '../store/user/user.slice';
 
 export const processErrorHandle = (message: string): void => {
   store.dispatch(setError(message));
