+++ src/store/user/user.api-actions.ts
@@ -1,57 +1,57 @@
-// import { createAsyncThunk } from '@reduxjs/toolkit';
-// import { AxiosInstance } from 'axios';
-// import { State } from 'history';
-// import { APIRoute, AppRoute, TIMEOUT_SHOW_ERROR } from '../../const';
-// import { dropToken, saveToken } from '../../services/token';
-// import { TAuthData } from '../../types/auth-data';
-// import { AppDispatch } from '../../types/state';
-// import { TUserData } from '../../types/user-data';
-// import { redirectToRoute } from '../actions';
-// import { setError } from './user.slice';
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { AxiosInstance } from 'axios';
+import { State } from 'history';
+import { APIRoute, AppRoute, TIMEOUT_SHOW_ERROR } from '../../const';
+import { dropToken, saveToken } from '../../services/token';
+import { TAuthData } from '../../types/auth-data';
+import { AppDispatch } from '../../types/state';
+import { TUserData } from '../../types/user-data';
+import { redirectToRoute } from '../actions';
+import { setError } from './user.slice';
 
-// export const clearErrorAction = createAsyncThunk(
-//   'USER/clearError',
-//   (_arg, { dispatch }) => {
-//     setTimeout(() => dispatch(setError(null)), TIMEOUT_SHOW_ERROR);
-//   }
-// );
+export const clearErrorAction = createAsyncThunk(
+  'USER/clearError',
+  (_arg, { dispatch }) => {
+    setTimeout(() => dispatch(setError(null)), TIMEOUT_SHOW_ERROR);
+  }
+);
 
-// export const checkAuthAction = createAsyncThunk<
-//   void,
-//   undefined,
-//   {
-//     dispatch: AppDispatch;
-//     state: State;
-//     extra: AxiosInstance;
-//   }
-// >('USER/checkAuth', async (_arg, { extra: api }) => {
-//   await api.get(APIRoute.Login);
-// });
+export const checkAuthAction = createAsyncThunk<
+  void,
+  undefined,
+  {
+    dispatch: AppDispatch;
+    state: State;
+    extra: AxiosInstance;
+  }
+>('USER/checkAuth', async (_arg, { extra: api }) => {
+  await api.get(APIRoute.Login);
+});
 
-// export const loginAction = createAsyncThunk<
-//   void,
-//   TAuthData,
-//   {
-//     dispatch: AppDispatch;
-//     state: State;
-//     extra: AxiosInstance;
-//   }
-// >(
-//   'USER/login',
-//   async ({ login: email, password }, { dispatch, extra: api }) => {
-//     const {
-//       data: { token },
-//     } = await api.post<TUserData>(APIRoute.Login, { email, password });
-//     saveToken(token);
-//     dispatch(redirectToRoute(AppRoute.Main));
-//   }
-// );
+export const loginAction = createAsyncThunk<
+  void,
+  TAuthData,
+  {
+    dispatch: AppDispatch;
+    state: State;
+    extra: AxiosInstance;
+  }
+>(
+  'USER/login',
+  async ({ login: email, password }, { dispatch, extra: api }) => {
+    const {
+      data: { token },
+    } = await api.post<TUserData>(APIRoute.Login, { email, password });
+    saveToken(token);
+    dispatch(redirectToRoute(AppRoute.Main));
+  }
+);
 
-// export const logoutAction = createAsyncThunk<
-//   void,
-//   undefined,
-//   { dispatch: AppDispatch; state: State; extra: AxiosInstance }
-// >('USER/logout', async (_arg, { extra: api }) => {
-//   await api.delete(APIRoute.Logout);
-//   dropToken();
-// });
+export const logoutAction = createAsyncThunk<
+  void,
+  undefined,
+  { dispatch: AppDispatch; state: State; extra: AxiosInstance }
+>('USER/logout', async (_arg, { extra: api }) => {
+  await api.delete(APIRoute.Logout);
+  dropToken();
+});
